requirements.txt
PyYAML==3.12
blessings==1.6
attrs==16.0.0
petname==2.0
urwid==1.3.1
juju==0.0.1
enforce==0.3.1
